+++ Sources/KeyboardKit/Layout/Providers/iPhoneKeyboardLayoutProvider.swift
@@ -204,7 +204,7 @@ struct iPhoneKeyboardLayoutProvider_Previews: PreviewProvider {
         return SystemKeyboard(
             layout: layout(for: locale).keyboardLayout(for: context),
             appearance: StandardKeyboardAppearance(context: context),
-            actionHandler: PreviewKeyboardActionHandler(),
+            actionHandler: .preview,
             width: context.previewWidth)
             .keyboardPreview(context: context)
             .background(previewImage, alignment: .bottom)