+++ Sources/KeyboardKit/Previews/UITextDocumentProxy+Preview.swift
@@ -8,9 +8,16 @@
 
 import UIKit
 
+public extension UITextDocumentProxy where Self == PreviewTextDocumentProxy {
+    
+    /**
+     This document proxy can be used in SwiftUI previews.
+     */
+    static var preview: UITextDocumentProxy { PreviewTextDocumentProxy() }
+}
+
 /**
- This class can be used to preview keyboard views. Don't use
- it in other situations.
+ This document proxy can be used in SwiftUI previews.
  */
 public class PreviewTextDocumentProxy: NSObject, UITextDocumentProxy {